projetos/data_engineering/credit_guard_ml_pipeline/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
ipykernel
requests
google-cloud-storage
pytz